Declare deadletter, requeue, and redeliver exchanges as Topic
Bucky declares the exchanges required for the requeue pattern as Direct type.
That means if the routing key used by a queue uses a wildcard, it will
not be able to route requeues or dead letters and they will be lost into
the ether :scream:

So the solution seems to be to use the Topic exchange type for these
exchanges. However this is quite a breaking change as existing exchanges
will have to be deleted to be redeclared. Not an easy one to roll out.